==> titanic/notebooks/3.0-ak-building-predictive-model.ipynb <==
## building predictive models

import pandas as pd
import os
import numpy as np

### import data

# set the path of the processed data
processed_data_path = os.path.join(os.path.pardir, 'data', 'processed')
train_file_path = os.path.join(processed_data_path, 'train.csv')
test_file_path = os.path.join(processed_data_path, 'test.csv')

train_df = pd.read_csv(train_file_path, index_col='PassengerId')
test_df = pd.read_csv(test_file_path, index_col='PassengerId')

train_df.info()

test_df.info()

### data preparation

X = train_df.loc[:, 'Age':].as_matrix().astype('float')
y = train_df['Survived'].ravel()

print X.shape, y.shape

# train test split
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=0)
print X_train.shape, y_train.shape
print X_test.shape, y_test.shape

# average survival in train and test
print 'mean survival in train: {0:.3f}'.format(np.mean(y_train))
print 'mean survival in test: {0:.3f}'.format(np.mean(y_test))

### check scikit-learn version

import sklearn

sklearn.__version__

#### make sure you have Scikit-learn v0.19 else update it and restart kernel

#!conda update -y scikit-learn

### baseline model

# import function
from sklearn.dummy import DummyClassifier

# create model
model_dummy = DummyClassifier(strategy='most_frequent', random_state=0)

# train model
model_dummy.fit(X_train, y_train)

print 'score for baseline model: {0:.2f}'.format(model_dummy.score(X_test, y_test))

# performance metrics
from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score

# accuracy score
print 'accuracy score for baseline model: {0:.2f}'.format(accuracy_score(y_test, model_dummy.predict(X_test)))

# confusion matrix
print 'confusion matrix for baseline model: \n {0}'.format(confusion_matrix(y_test, model_dummy.predict(X_test)))

# precision and recall scores
print 'precision score for baseline model: {0:.2f}'.format(precision_score(y_test, model_dummy.predict(X_test)))
print 'recall score for baseline model: {0:.2f}'.format(recall_score(y_test, model_dummy.predict(X_test)))

### first kaggle submission

# converting to the matrix
test_X = test_df.as_matrix().astype('float')

# get predictions
predictions = model_dummy.predict(test_X)

df_submission = pd.DataFrame({'PassengerId': test_df.index, 'Survived': predictions})

df_submission.head()

submission_data_path = os.path.join(os.path.pardir, 'data', 'external')
submission_file_path = os.path.join(submission_data_path, '01_dummy.csv')

df_submission.to_csv(submission_file_path, index=False)

def get_submission_file(model, filename):
    # converting to the matrix
    test_X = test_df.as_matrix().astype('float')
    # make predictions
    predictions = model.predict(test_X)
    # submission dataframe
    df_submission = pd.DataFrame({'PassengerId': test_df.index, 'Survived': predictions})
    # submission file
    submission_data_path = os.path.join(os.path.pardir, 'data', 'external')
    submission_file_path = os.path.join(submission_data_path, filename)
    # write to the file
    df_submission.to_csv(submission_file_path, index=False)

# get submission file
get_submission_file(model_dummy, '01_dummy.csv')

### logistic regression model

# import function
from sklearn.linear_model import LogisticRegression

# create model
model_lr_1 = LogisticRegression(random_state=0)

# train model
model_lr_1.fit(X_train, y_train)

# evaluate model
print 'score for logistic regression - version 1: {0:.2f}'.format(model_lr_1.score(X_test, y_test))

# performance metrics
# accuracy
print 'accuracy for logistic regression - version 1: {0:.2f}'.format(accuracy_score(y_test, model_lr_1.predict(X_test)))
# confusion matrix
print 'confusion matrix for logistic regression - version 1: /n {0}'.format(confusion_matrix(y_test, model_lr_1.predict(X_test)))
# precision
print 'precision for logistic regression - version 1: {0:.2f}'.format(precision_score(y_test, model_lr_1.predict(X_test)))
# recall
print 'recall for logistic regression - version 1: {0:.2f}'.format(recall_score(y_test, model_lr_1.predict(X_test)))

# model coefficients
model_lr_1.coef_

### second kaggle submission

# get submission file
get_submission_file(model_lr_1, '02_lr.csv')

## part 2

### hyperparameter optimization

# base model
model_lr = LogisticRegression()

from sklearn.model_selection import GridSearchCV

parameters = {'C':[1.0, 10.0, 50.0, 100.0, 1000.0], 'penalty':['l1', 'l2']}
clf = GridSearchCV(model_lr, param_grid=parameters, cv=3)

clf.fit(X_train, y_train)

clf.best_params_

print 'best score: {0:.2f}'.format(clf.best_score_)

# evaluate model
print 'score for logistic regression - version 2: {0:.2f}'.format(clf.score(X_test, y_test))

### making third submission

# get submission file
get_submission_file(clf, '03_lr.csv')

### feature normalization and standardization

from sklearn.preprocessing import MinMaxScaler, StandardScaler

#### feature normalization

# feature normalization
scaler = MinMaxScaler()
X_train_scaled = scaler.fit_transform(X_train)

X_train_scaled[:,0].min(),X_train_scaled[:,0].max()

# normalize test data
X_test_scaled = scaler.transform(X_test)

#### feature standardization

# feature standardization
scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

#### create model after standardization
model_lr = LogisticRegression()
parameters = {'C':[1.0, 10.0, 50.0, 100.0, 1000.0], 'penalty':['l1', 'l2']}
clf = GridSearchCV(model_lr, param_grid=parameters, cv=3)
clf.fit(X_train_scaled, y_train)

clf.best_score_

# evaluate model
print 'score for logistic regression - version 2: {0:.2f}'.format(clf.score(X_test_scaled, y_test))

### model persistence

# import pickle library
import pickle

# create file path
model_file_path = os.path.join(os.path.pardir, 'models','lr_model.pkl')
scaler_file_path = os.path.join(os.path.pardir, 'models','lr_scaler.pkl')

# open the files to write
model_file_pickle = open(model_file_path, 'wb')
scaler_file_pickle = open(scaler_file_path, 'wb')

# persist the model and scaler
pickle.dump(clf, model_file_pickle)
pickle.dump(scaler, scaler_file_pickle)

# close the file
model_file_pickle.close()
scaler_file_pickle.close()

#### load the persisted file

# open files in read mode
model_file_pickle = open(model_file_path, 'r')
scaler_file_pickle = open(scaler_file_path, 'r')
# load files
clf_loaded = pickle.load(model_file_pickle)
scaler_loaded = pickle.load(scaler_file_pickle)
# close files
model_file_pickle.close()
scaler_file_pickle.close()

clf_loaded

scaler_loaded

# transform the test data using loaded scaler object
X_test_scaled = scaler_loaded.transform(X_test)
# calculate the score used loaded model object
print 'score for persisted logistic regression: {0:.2f}'.format(clf_loaded.score(X_test_scaled, y_test))
